Create URL shortener
To Design a system that takes big URLs and converts them into a short 6 character URL. 
It is given that URLs are stored in the database and every URL has an associated integer id. 
One important thing to note is, the long URL should also be uniquely identifiable from the short URL.

One Simple Solution could be Hashing. Use a hash function to convert long string to short string.
 In hashing, that may be collisions (2 long URLs map to same short URL) and we need a unique short URL for every long URL so that we can access long URL back.
A Better Solution is to use the integer id stored in the database and convert the integer to a character string that is at most 6 characters long. 
This problem can basically seen as a base conversion problem where we have a 10 digit input number and we want to convert it into a 6 character long string.